requirements.txt
fastapi==0.128.8
gdown==5.2.1
numpy==2.4.2
pandas==3.0.0
Pillow==12.1.1
pydantic==2.12.5
Requests==2.32.5
streamlit==1.54.0
tensorflow==2.19.0
